feat: Multiple enhancements and bug fixes across admin panel, settings, and activity logs
Backend Changes:
- Activity Log: Fixed student name search functionality in Activity Log API
  * Corrected MongoDB query structure for user search filter
  * Properly combined regex search with admin exclusion filter
  * Handles existing  arrays from skill filters correctly

- Activity Log: Improved time ago display formatting
  * Fixed singular/plural forms (1 day vs 1 days, 1 hour vs 1 hours)
  * Added proper grammar for minutes, hours, days, and months

- Suggestions: Implemented persistent dismissed suggestions
  * Added dismissedSuggestions field to User model
  * Created POST /api/suggestions/dismiss endpoint
  * Updated GET /api/suggestions to filter dismissed suggestions
  * Dismissed suggestions now persist across page refreshes

Frontend Changes:
- Admin Panel - Activity Log:
  * Fixed student name search functionality
  * Improved time ago display with proper singular/plural forms
  * Enhanced user experience with better date/time formatting

- Admin Panel - Settings:
  * Removed 'Delete Account' functionality from Danger Zone
  * Removed handleDeleteAccount function
  * Cleaned up unused code and imports

- User Settings Page:
  * Removed 'Delete Account' button and Danger Zone section
  * Removed unused Trash2 icon import
  * Removed handleDeleteAccount function

- Suggestions Page:
  * Implemented persistent dismissed suggestions
  * Added API integration for dismissing suggestions
  * Dismissed suggestions now persist across page refreshes
  * Added proper error handling with UI rollback on API failures

Other Improvements:
- Enhanced password validation (uppercase/lowercase requirements)
- Improved role history performance score decimal handling
- Better error handling and user feedback throughout
- Code cleanup and optimization

+++ backend/routes/roleHistory.js
@@ -96,7 +96,7 @@ router.post("/", protect, async (req, res) => {
       date: date.trim(),
       team: teamString,
       skills: skills || [],
-      score: parseInt(performance),
+      score: parseFloat(performance),
       description: notes || "",
       imageUrl: imageUrl || "",
       projectImages: projectImages || [],
@@ -183,7 +183,7 @@ router.patch("/:projectId", protect, async (req, res) => {
         if (date !== undefined) project.date = date.trim();
         if (teamMembers !== undefined) project.team = teamString;
         if (skills !== undefined) project.skills = skills;
-        if (performance !== undefined) project.score = parseInt(performance);
+        if (performance !== undefined) project.score = parseFloat(performance);
         if (notes !== undefined) project.description = notes;
         if (imageUrl !== undefined) project.imageUrl = imageUrl.trim();
         if (projectImages !== undefined) project.projectImages = projectImages;
